UV Selection + Cleanup
UV map can now be selected as opposed to it being hardcoded at index 1.
Added a warning if the selected bake map does not exist on the active object.
A warning also appears if there's no lamp that casts shadows.
Allowed passing a name and bake_margin.
The generated bake map is now selected in the UV image editor.